# Week11/3.Wednesday/codegen/src/gen/src/main/java/org/openapitools/api/ApiApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Book;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-12-20T16:53:13.725632200-05:00[America/New_York]")
@Validated
@Tag(name = "book-controller", description = "the book-controller API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /api/books
     *
     * @param book  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createBook",
        tags = { "book-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/books",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Book> createBook(
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"price\" : 1.4658129805029452, \"author\" : { \"name\" : \"name\", \"id\" : 5 }, \"genre\" : \"FICTION\", \"id\" : 0, \"publishedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"title\" : \"title\", \"likes\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /api/books/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteBook",
        tags = { "book-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/books/{id}"
    )
    
    default ResponseEntity<Void> deleteBook(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/books
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getAllBooks",
        tags = { "book-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", array = @ArraySchema(schema = @Schema(implementation = Book.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/books",
        produces = { "*/*" }
    )
    
    default ResponseEntity<List<Book>> getAllBooks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "[ { \"price\" : 1.4658129805029452, \"author\" : { \"name\" : \"name\", \"id\" : 5 }, \"genre\" : \"FICTION\", \"id\" : 0, \"publishedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"title\" : \"title\", \"likes\" : 6 }, { \"price\" : 1.4658129805029452, \"author\" : { \"name\" : \"name\", \"id\" : 5 }, \"genre\" : \"FICTION\", \"id\" : 0, \"publishedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"title\" : \"title\", \"likes\" : 6 } ]";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/books/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getBookById",
        tags = { "book-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/books/{id}",
        produces = { "*/*" }
    )
    
    default ResponseEntity<Book> getBookById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"price\" : 1.4658129805029452, \"author\" : { \"name\" : \"name\", \"id\" : 5 }, \"genre\" : \"FICTION\", \"id\" : 0, \"publishedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"title\" : \"title\", \"likes\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /api/books/{id}
     *
     * @param id  (required)
     * @param book  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateBook",
        tags = { "book-controller" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "*/*", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/books/{id}",
        produces = { "*/*" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Book> updateBook(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("*/*"))) {
                    String exampleString = "{ \"price\" : 1.4658129805029452, \"author\" : { \"name\" : \"name\", \"id\" : 5 }, \"genre\" : \"FICTION\", \"id\" : 0, \"publishedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"title\" : \"title\", \"likes\" : 6 }";
                    ApiUtil.setExampleResponse(request, "*/*", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}